update for python3

+++ adas.py
@@ -1,9 +1,10 @@
 import os
 import errno
 import shutil
-import urllib
+import urllib.request, urllib.parse, urllib.error
+import ssl
 
-open_adas = 'http://open.adas.ac.uk/'
+open_adas = 'https://open.adas.ac.uk/'
 
 class OpenAdas(object):
     def search_adf11(self, element, year='', ms='metastable_unresolved'):
@@ -30,13 +31,13 @@ class OpenAdas(object):
         else:
             __, path = url_filename
 
-        tmpfile, __ = urllib.urlretrieve(url)
-
         dst_filename = os.path.join(self.dst_directory, path)
-        self._mkdir_p(os.path.dirname(dst_filename))
+        if not os.path.exists(dst_filename):
+            tmpfile, __ = urllib.request.urlretrieve(url)
 
+            self._mkdir_p(os.path.dirname(dst_filename))
 
-        shutil.move(tmpfile, dst_filename)
+            shutil.move(tmpfile, dst_filename)
 
     def _construct_url(self, url_filename):
         """
@@ -86,8 +87,9 @@ class AdasSearch(object):
         return self._parse_data()
 
     def _retrieve_search_page(self):
-        search_url =  self.url + urllib.urlencode(self.parameters)
-        res, __ = urllib.urlretrieve(search_url)
+        search_url =  self.url + urllib.parse.urlencode(self.parameters)
+        ssl._create_default_https_context = ssl._create_unverified_context
+        res, __ = urllib.request.urlretrieve(search_url)
         self.data = open(res).read()
         os.remove(res)
 
@@ -119,12 +121,11 @@ class AdasSearch(object):
         return int(id_)
 
 
-from HTMLParser import HTMLParser
+from html.parser import HTMLParser
 class SearchPageParser(HTMLParser):
     """
     Filling in a search form on http://open.adas.ac.uk generates a HTML document
     with a table that has the following structure:
-
     >>> html = '''
     ... <table summary='Search Results'>
     ...     <tr>
@@ -134,7 +135,6 @@ class SearchPageParser(HTMLParser):
     ...     <td>C</td> <td><a href='filedetail.php?id=32154'>rc89_c.dat</a></td>
     ...     </tr>
     ... </table>'''
-
     The SearchPageParser can parse this document looking for a table with a
     class `searchresults`.
     >>> parser = SearchPageParser()
